Get rid of and refactoring old code
