+++ src/main/java/api/gradeAPI_for_reference/GradeDB.java
@@ -1,8 +1,8 @@
 // FOR REFERENCE ONLY, DO NOT USE IN PRODUCTION (managed by Henry)
-package api;
+package api.gradeAPI_for_reference;
 
-import entity.gradeAPI_for_reference.Grade;
-import entity.gradeAPI_for_reference.Team;
+import entity.gradeAPI_entities_for_reference.Grade;
+import entity.gradeAPI_entities_for_reference.Team;
 import org.json.JSONException;
 
 public interface GradeDB {
