--- src/main/java/api/jsoup.java
@@ -1,23 +1,0 @@
-// Author: Bora
-
-//package api;
-//
-//import org.jsoup.Jsoup;
-//import org.jsoup.nodes.Document;
-//
-//
-//import java.io.IOException;
-//
-//public class jsoup {
-//    public static void main(String[] args) {
-//        Document doc = null;
-//        try {
-//            doc = Jsoup.connect("https://en.wikipedia.org/").get();
-//        } catch (IOException e) {
-//            throw new RuntimeException(e);
-//        }
-//        System.out.println(doc.outerHtml());
-//
-//
-//    }
-//}
